Updates for NIM credentials view:
1. Credentials and expanded headers that show countdown clocks (time left) need to be redrawn periodically.  This is done by setting a timer for a row that fires when the clock on that that row needs to be redrawn.  When switching between the basic and advanced view, the timers need to the discarded because the rows and clocks that are displayed will change.  However, if a timer fires just as we are switching modes, the associated WM_TIMER message will be placed on the message queue and will not be removed when we try to kill the timer using KillTimer().  Therefore, we might receive a WM_TIMER message that corresponds to a row that no longer exists, or has changed.

Earlier, the code tried to assert that WM_TIMER messages we receive were consistent with current data.  Since receiving an inconsistent WM_TIMER message is not an error, this patch adds code for verifying that the timer event we have received is consistent with the data, and if not, quietly ignore the message instead of failing an assert.

2. Fix a typo.  The third parameter to cw_select_row() is actually a WPARAM value containing a bit field corresponding to the modifier keys that were in effect while the selection was made.

3. The user can invoke the context menu using the keyboard.  If they do, then the x and y co-ordinates provided with the WM_CONTEXTMENU message will be set to (-1, -1).  Add code to check for this situation and use the current cursor_row value as the target of the context menu.

4. If the user right-clicks on a row that is not part of the current selection, the earlier code would set the cursor context to the row that the user right-clicked on while leaving the selection as is.  Since some actions (such as 'new credentials' and 'properties') use the cursor context and some other actions (such as 'destroy credentials') use the selection context, this results in inconsistent behavior.

This patch adds code that update the selection to include the row that the user right-clicked on if that row is not part of the selection.  Doing so would provide a consistent experience where actions that operate on the cursor context and actions that operate on the selection both operate on the same credentials that the user sees as selected.

5.  Remove unused code.