Make font changes be reflected immediately in the UI
Due to the changes in how drawing is handled in NIM, changes to fonts in the appearance configuration panel were not being immediately reflected in the UI.

As a side-effect, we no longer cache fonts in the resource cache and use them directly.  The only consumers for GDI objects like that should be the functions in drawfuncs.h.