Fix context menu default action behavior
If the default action for the notification icon context menu is
not KHUI_ACTION_NEW_CRED, we couldn't mark the default action on
the context menu because this action is no longer present on the
context menu.  It had been replaced by KHUI_MENU_NEW_CRED.

To fix this, remove the first section of menu items off the
context menu and insert the default action and, if necessary, the
application open/close action. Then mark the default action as
default.