New build setup script
The old build setup script was specific to my machine and had to
be modified each time I changed machines or even build
directories.  With this script, hopefully we will be able to use
a single master build setup script.

Any machine specific setup is now done in setupbuild.local.cmd
which is called from setupbuild.cmd.  The arguments to
setupbuild.cmd are the same as the arguments to the Platform SDK
build environment.  All the build type values are derived from
the variables setup by the Platform SDK SetEnv.Cmd.

We support Platform SDK version v6.1 for now.