khui_menu_dup() shouldn't try to associate duplicate menu with command
A menu in NIM can be associated with a default action.  Only one menu
can be associated with a command.  When duplicating a menu, the newly
created menu should not be associated with the original command.  This
is how khui_menu_dup() was documented to work, but it was attempting
to call khui_menu_create() with the original command.

Change-Id: I58f563f0092350b021864a603d9e65e3c660d0cb