Introduce context menu notifications
Plug-ins will now be able to customize context menus that are going to
be displayed for credentials window object before they are displayed.
This is accomplished by sending a KMSG_CRED_CTX_MENU notification
prior to the menu being displayed.  Each KMSG_CRED subscriber will be
given an opportunity to use the khui_context_menu_*() functions to
access and modify the menu.

Change-Id: Idfc21b0c9e0acfb07fd11e214488da24360914e7