Don't leak context references when enumerating
The kherr_get_next_context() function was not releasing the handle to
the error context that was being passed in.  The only consumer of this
API was operating under the assumption that it was.

Technically this is an API change since the fact that the handle was
being released was not documented.  Now it is documented.

Change-Id: Ia415d69daeaef5a692ef6346613f558aee83f8b6