Use khm_int32 instead of time_t in khui_tracker
khui_tracker used time_t as the data type for keeping track of time
intervals in seconds.  The time_t type can be 32-bit or 64-bit on
Windows depending in compiler options regardless of the target
processor architecture, which caused incompatibilities with plug-ins.

We are using khm_int32, which is always a 32-bit quantity regardless
of the compiler options used.  Since all the distributed binaries were
using 32-bit time_t values, this ensures that the generated binaries
are backwards compatible.

Change-Id: I31d508d94fa5630f19d4b5dd8fbbb4be52497b7a