KeyStore: Encode/decode optional headers
The KeyStore encoding includes the ability to specify optional
fields in a manner that preserves format compatibility with
previous versions of the codec.  It does this by allowing older
codecs to skip over fields that were introduced later and read in
only those fields that it knows about.

The logic for skipping over unrecognized optional fields had a
bug where it would skip over any adjacent optional fields that
didn't belong to the current object. I.e.:

Obj1
  Field1_1
  Field1_2
  Obj2
   Field2_1
   Field2_2
   OptionalField2_3 <--
  OptionalField1_3

In the above object, the old logic for skipping unrecognized
optional fields, when invokend at the marked position of the
encoding, would skip over OptionalField1_3 because that field is
not recognized by the Obj2 decoder.  The correct behavior
introduced in this patch would stop the skipping operation when
it reached the end of Obj2 leaving OptionalField1_3 to be dealt
by the decoder for Obj1.

Change-Id: Id692f0f4426869418977e7c1f9b5db8c7bdb5a61