Define __cplusplus when building documentation
The API now includes classes that can be used by C++ programs.  They
are wrapped in __cplusplus conditionals to prevent them from breaking
C builds.  For the documentation build, Doxygen should be reading them
instead of skipping over them.  Define the __cplusplus macro in the
configuration to make this happen.

Change-Id: I22c988e3c4e028cb231ba3c128cd7591c0a48cc2