UI improvements for identity icon selection dialog
The identity icon selection dialog used to make HTTP requests from the
UI thread.  Avoid doing that and use the new asynchronous HTTP request
class defined in HttpRequest.

In addition, display user-friendly error messages when the requested
resources are not found.  Also, allow the user to cancel out of the
operation if it is taking too long.

Change-Id: Id02f972a9393679c4ee02b7e431d0e8eef99adca