Krb5: Synchronize credentials provider and identity provider
This is a catch-up to an architectural change that was made to the NIM
v2 credentials and identity providers.  The problem meant to be solved
is that if there is a credentials providers and an identity provider
that are tightly bound (such is the case with Krb5Cred/Krb5Ident, as
well as others), then they need to synchronize start-up with each
other.

In the early versions of NIM, we had the credentials provider depend
on the identity provider.  However this lead to problems during
startup where the identity provider couldn't really function with-out
the credentials provider already running (i.e. can't provide a default
identity etc.).  Therefore, the dependency was reversed.  So now the
identity providers depend on the credentials providers.  However,
there is still a dependency issue where the credentials providers
can't usually create report any credential until after the identity
provider has started.  Therefore, we use the following scheme to
synchronize providers:

1. Cred-provider handles INIT for plug-in (no dependencies)

   While handling INIT, it creates an event object that is not
   signalled by default.

2. Ident-provider handles INIT for plug-in (depends on Cred-provider)

   Since there is a dependency on Cred-provider, this step is
   guaranteed by the NIM runtime to happen after step 1.

3. Ident-provider handles IDENT_INIT for identity provider

   Guaranteed to happen after step 2, and by extension after step 1.
   Therefore, the event object has already been initialized.  While
   handling IDENT_INIT, the Ident-provider signalls the event object.

3. Cred-provider starts message loop

   The message handler for CRED messages block on the event object
   created in step 1.

   Regardless of when the Ident-provider handles its INIT message,
   this step will always happen after step 1 because this happens in
   the same plug-in thread following the Cred-provider INIT
   processing.  Therefore, if we get here before the Ident-provider
   does INIT processing, then we have to wait until the Ident-provider
   handles IDENT_INIT and signalls this event.

   By the time the event is signalled, both the Cred-provider and
   Ident-provider are ready for operation and it is safe to handle
   CRED messages, create credentials and create identities.

Change-Id: I28f1ad44a0f10f2a19ea7b9a44b33f621a6078f6