Don't generate a redundant change notification when setting identity parent
If the new parent identity is the same as the old one, don't generate
a change notification.

In addition, identity parent change notifications should always be
sent out even when the "no notify" bit is on.  The
KCDB_IDENT_FLAG_NO_NOTIFY bit only applies to KCDB_OP_MODIFY
notifications.

Change-Id: I388db02c84c0ee6b927a3daf555b7aafecde30b4