Don't collapse outlines with no children
Credentials display outlines that have no children were being
automatically collapsed even though doing so is not necessary.  The
outline widget is only shown if there are children, so collapsing
doesn't have any visible effect.

Once an outline is collapsed, it retains its collapsed state even if
child outlines are added at a later time.  E.g.: a pinned identity
with no credentials will be automatically collapsed, but will remain
collapsed when new credentials are obtained for it.  This seems to
counter expected behavior.  Therefore, we now retain the expanded
state of the outline even when there are no children.

Change-Id: Iebe588726c512bff3863a6b4475ddda8c839175f