Use SetErrorMode() to suppress errors with LoadLibrary()
We use LoadLibrary() to load plug-ins.  If there are unresolved
imports or if a dependent DLL cannot be found, Windows will show an
error indicating such.  The call to SetErrorMode() suppresses this
dialog and instead returns a failure code from LoadLibrary().  This is
desirable, since we are already handling load failures gracefully and
can prevent unnecessary interruptions when NetIDMgr is being loaded at
startup.

Change-Id: If79a83465e373cd05902390c00f383dfdeb35f8d