Support Omaha Update
Add support for registering as a client with the Secure Endpoints
Omaha Update Service.  If Omaha Update is not running on the machine,
then the extra registry entries are harmless and will be removed with
the application.  If the Omaha Update service is running, then
registering as a client will cause Network Identity Manager to be
automatically kept up-to-date.

Change-Id: If082decddb772d58d6b2905f0aaee34f89a8dc14