KeyStore: Complain of password mismatches proactively
When changing the KeyStore password, the user is presented with three
text boxes:

   - Password          [ ****                         ]
   - New password      [ ****                         ]
   - Confirm password  [ ****                         ]

For the user to proceed, the `new password` and the `confirm password`
values must be the same.  If the user mistypes one, it won't be very
apparent at this stage unless the password lengths are different
(since we mask the password characters) and the user would not know
why she is not being shown the 'Next >>' or 'Finish' button.

This patch modifies the dialog code to show a hint to the user when
the `new password` and `confirm password` values are different.

Change-Id: I1c627e2ca0354a93e8fb8feee4b7bb0c386ec38f