CertProv: Temporarily set initializable flag for all identities
Not setting the flag breaks backwards compatilibity for some third
party plug-ins.  This isn't a production plug-in, so introduce a
temporary hack so that people who use this code can continue testing.
Eventually, we will only be setting the flag for certificates that can
be used to initialize an identity.

Change-Id: Ied97bf0e7a9fb0b9f235093943d6dce33570452a