kherr: Thread safe notifications
kherr context event notifications were not thread safe.  When a
context event occurs (e.g. a debug event), the notification indicates
the type of event (e.g. that it is a debug event), but doesn't include
the event data.  When the event handler intercepts the event, it
doesn't have a method to accurately determine the data associated with
the event (e.g. the message that was reported).  It can query the
context for the last such event, but there is no guarantee that the
last such event is the one that is associated with this specific
notification.

This patch adds the context data to the notification.

Change-Id: I3bb0fc87ad190f7a3ab491ac33d58e54a89eda5d