Support install-local for krb4 and krb5 plug-ins
The install-local feature was added to plug-in sample code as an aid
during development.  The install-local target of the plug-in Makefile
adds the necessary registry keys to the user's registry hive to load
and execute the plug-in binaries off the build tree.

Add this capability to the MITKrb4 and MITKrb5 plug-ins.

Change-Id: Ic47267a2f020aef50a1e3f4b6669972bb37a1b33