Support install-local for krb4 and krb5 plug-ins
The install-local feature was added to plug-in sample code as an aid
during development.  The install-local target of the plug-in Makefile
adds the necessary registry keys to the user's registry hive to load
and execute the plug-in binaries off the build tree.

Add this capability to the MITKrb4 and MITKrb5 plug-ins.

Change-Id: I2ae72070916ecb5afb69e6b53cf36114a1849e82