Build loadfuncs.c within the NIM source tree
Linking against a static library that contains manifest dependency
directives is not safe.  The loadfuncs.lib that is distributed with
some versions of KfW contains manifest directives. If the version of
Visual C used to build NIM is not the exact same version as the
version used to build loadfuncs.lib, the resulting binaries will
depend on both versions of the Visual C runtime library.  Since we
only ship one version of the runtime, the binaries will not
successfully load on a clean target machine.

Therefore, this patch includes loadfuncs.c in plugins/common so that
it can be built as part of the NIM build.

Change-Id: I839d17ba2b9a14595bc848d6ff5ff7c31c254594