Add internal API khui_cw_set_primary_id_no_notify()
An API is required to change the primary identity for a new
credentials operation that does not perform a notification.  This is
used to replace incorrect logic used within the new credentials wizard
to switch the primary identity.

A notification is not required because the switch is performed by the
new credentials wizard itself.

Change-Id: I558e0e6484e2b59340ad9244dac0f931a6697dc1