Change internal API for listening for new credentials dialog results
The previous internal API only returned the name of the primary
identity.  For new credentials operations that involve KeyStores and
other complicated configurations, this is insufficient.  Therefore the
new API uses a callback notification when the new credentials
operation completes.  The callback can examine the khui_new_creds
object as necessary.

Change-Id: I88c12130c807aad5cfab3138fa8652992ed0eda3