Remember target identity when switching identities
The NewCredWizard class examines the primary identity and changes it
if the primary identity has a parent.  KeyStores that contain keys for
a derived identity are indicated as being parents of the derived
identity.  When new credentials are being obtained for a derived
identity, this mechanism switches the primary identity to the KeyStore
so that only the KeyStore password is required.

When switching, add the original identity as an additional identity so
that the original target can be tracked.  Otherwise, when the new
primary identity (KeyStore) has several derived identities, there
would be no way to track which one was the original target.

Change-Id: Id91715046017320cc494de1f4e5d259d53ee50b1