Proper determination of credentials cache for remote requests
When remote requests are received (such as from GSSAPI) for new
credentials, we return the resulting credentials cache name to the
caller.  When the new credentials operation involved a KeyStore or
some other non-straightforward configuration, then the primary
identity for the new credentials operation alone is insufficient to
determine the credentials cache.  The primary identity in this case
would normally be the KeyStore.  Plus, some of the other identities
may not be Kerberos v5.

Use the new dialog completion notification API to go through the list
of identities that were involved in the new credentials operation and
then determine which one to pick the credentials cache from.  We
choose the first Kerberos v5 identity.

When a Kerberos v5 identity is selected for a new credentials
operation, but the NewCredWizard switches the primary identity because
the the selected identity was associated with a KeyStore, the original
identity is added as the next identity in the list.  Therefore,
selecting the first Kerberos v5 identity will:

1) If the primary identity was a KeyStore, the first Kerberos v5
   identity is the original identity for whom we requested credentials
   for.

2) If we didn't start from a Kerberos v5 identity (i.e. had to go
   through an identity specification dialog), then the first Kerberos
   v5 identity will not be significant, but we don't have any more
   information to pick a better one.

Change-Id: I154c44e2af8f46456feaad9d5035667658964266