Allow use of weak encryption types
A new per-identity parameter 'allow_weak_crypto' can be used to
control whether weak encryption types are enabled for an
identity.  If it is, then the following encryption types will be
enabled prior to obtaining new credentials for that identity
(Kerberos v5 only):

 ETYPE_DES_CBC_CRC
 ETYPE_DES_CBC_MD4
 ETYPE_DES_CBC_MD5
 ETYPE_DES_CBC_NONE
 ETYPE_DES_CFB64_NONE
 ETYPE_DES_PCBC_NONE

Due to limitations in the underlying Kerberos implementation, the
change is global to the instance of Network Identity Manager.

Change-Id: I0e97e4300af6fc9e8da7e2bcf821e995621f036f