Help docs updated with removal of Krb4 support
With NIM 2.1, support for Kerberos v4 has been deprecated and
the provider has been removed as part of the transition to
Heimdal. The html help files have been updated to reflect the
removal of the functionality in baseline features, what's new,
note highlights on relevant pages and removal of links to
Kerberos v4 specific pages. v4 pages have been left if in the
future an optional v4 package is reintroduced due to customer
need.

Change-Id: Id9a6b02d8b3c58e1c3011e927e73c43eb78981f5