Kerberos 5: Don't participate in non-Krb5 renewals
The Kerberos v5 plug-in should check whether the primary identity
of a renewal operation is a Kerberos identity.  If it isn't, it
doesn't participate in the operation and disables itself.  This
signals to other providers that depend on Kerberos v5 that they
shouldn't participate in the renewal either.

This signal is necessary for compatibility with third party NIM
1.x plug-ins.

Change-Id: Ic56331ba6e8423a24474d12073bcbe8338733de9