Workaround infinite cpu usage in kherr.c get_progress()
The source of the error has not been identified yet but it is
possible to prevent the 100% cpu utilization that is a side effect.
It has been noticed several times (and captured under a debugger)
that cc == TNEXTSIBLING(cc) == TPREVSIBLING(cc).  This results in
the for loop never completing.

If such a condition is noticed, fix up the cc->prev and cc->next
pointers to refer to NULL and in DEBUG builds throw an assert().

Change-Id: I96192bd62ecdc8b943eef1b6c7b669ee267892f0